Shopping-App
Application contain 3 screen
Home screen : -
A home page where all products are displayed in a list. With every product, there is a count (default 0) which signifies the number of that product in the cart. The home page also have a dropdown option from which you can select a category. Clicking selecting a category filter the items on the home page to be only from that specific category (instead of all products). The product item also have a “Like” button. On pressing the Like button, the product’s category stored as a “liked category” with use of shared prefrences. The Home page have a “Suggested categories” option. On selecting this, the products of the categories stored as “liked category” displayed.
Product detail screen :-
In these screen details of product
Also have to buttons addtocart and remove button addtocart update the count of itemes in cart nd remove button to remove all items and also checkout button to land on checkout page
Checkout Screen :-
it contain list of all product added in cart and total amount and each item contain to removeitem from cart and all buy button to restart activity and set cart count to 0
sqlite dbhelper used to store product title and product count in cart after buy button datat base is cleared

in product object likedstatus is used to set image of imageview of like button not to filter liked categories